refactor: rm redux
