+++ react.ts
@@ -1,4 +1,3 @@
-export * from "./query/react.ts";
 export * from "./store/react.ts";
-export { Provider, useDispatch, useSelector } from "./deps.ts";
+export { useDispatch, useSelector } from "./deps.ts";
 export type { TypedUseSelectorHook } from "./deps.ts";
