--- redux/types.ts
@@ -1,6 +1,0 @@
-import type { AnyAction } from "../types.ts";
-
-export interface StoreLike<S = unknown> {
-  getState: () => S;
-  dispatch: (action: AnyAction) => void;
-}
