+++ store/types.ts
@@ -36,6 +36,18 @@ export type Output<O extends { [key: string]: BaseSchema<unknown> }> = {
   [key in keyof O]: O[key]["initialState"];
 };
 
+export interface FxMap {
+  loaders: <M extends AnyState>(s: string) => LoaderOutput<M, AnyState>;
+  cache: (s: string) => TableOutput<any, AnyState>;
+  [key: string]: (name: string) => BaseSchema<unknown>;
+}
+
+export interface FxSchema<S extends AnyState, O extends FxMap = FxMap> {
+  db: { [key in keyof O]: ReturnType<O[key]> };
+  initialState: S;
+  update: FxStore<S>["update"];
+}
+
 export interface FxStore<S extends AnyState> {
   getScope: () => Scope;
   getState: () => S;
